refactor: likes 테이블 이름 조회에 있어 Entity이름과 테이블 이름이 다르기 때문에 그 문제를 해결하기 위해 이름 수정.
