+++ src/main/java/com/hotnerds/post/domain/like/Likes.java
@@ -1,63 +1,49 @@
 package com.hotnerds.post.domain.like;
 
-import com.hotnerds.common.exception.BusinessException;
-import com.hotnerds.common.exception.ErrorCode;
+import com.hotnerds.post.domain.Post;
 import com.hotnerds.user.domain.User;
 import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Embeddable;
-import javax.persistence.FetchType;
-import javax.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
+import javax.persistence.*;
+import java.util.Objects;
 
-@AllArgsConstructor
-@Embeddable
+@Entity
+@Getter
+@Table(name = "LIKES")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Likes {
-    @OneToMany(
-            mappedBy = "post",
-            fetch = FetchType.LAZY,
-            cascade = CascadeType.PERSIST,
-            orphanRemoval = true
-    )
-    private List<Like> likes = new ArrayList<>();
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
 
-    public int getCount() {
-        return likes.size();
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id")
+    private Post post;
+
+    @Builder
+    public Likes(Long id, User user, Post post) {
+        this.id = id;
+        this.user = user;
+        this.post = post;
     }
 
-    public boolean contains(Like like) {
-        return likes.stream()
-                .anyMatch(like::equals);
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Likes like = (Likes) o;
+        return Objects.equals(user, like.user) && Objects.equals(post, like.post);
     }
 
-    public void add(Like like) {
-        if(likes.contains(like)) {
-            throw new BusinessException(ErrorCode.DUPLICATED_LIKE_EXCEPTION);
-        }
-        likes.add(like);
-    }
-
-    public void remove(Like like) {
-        if(!likes.contains(like)) {
-            throw new BusinessException(ErrorCode.LIKE_NOT_FOUND_EXCEPTION);
-        }
-        likes.remove(like);
-    }
-
-    public List<User> getLikesUsers() {
-        return likes.stream()
-                .map(like -> like.getUser())
-                .collect(Collectors.toList());
-    }
-
-    public static Likes empty() {
-        return new Likes(new ArrayList<>());
+    @Override
+    public int hashCode() {
+        return Objects.hash(user, post);
     }
 }
